requirements.txt
Flask
Flask-PyMongo
python-dotenv
email-validator
flask-mail
python-dateutil
matplotlib
numpy